pyproject.toml
[project]
name = "kissflow-downloader"
version = "0.1.0"
description = "Kissflowのワークフローから添付ファイルをダウンロードするツール"
requires-python = ">=3.13.3"
dependencies = [
    "selenium>=4.18.1",
    "webdriver-manager>=4.0.1",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["kissflow_downloader"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = ["selenium.*", "webdriver_manager.*"]
ignore_missing_imports = true
